ipam: Move pkg/ipam/allocator/operator to pkg/ipam/allocator/clusterpool
[ upstream commit e613b5947db26453cc7f1af246a527316da10aac ]
